Group loops and generics tests
