--- tests/ForDownTo.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class ForDownToDemo {
-        public static void Test() {
-            for (var i = 3; i >= 1; i--) System.Console.WriteLine(i);
-        }
-    }
-}
