--- tests/ForEach.cs
@@ -1,10 +1,0 @@
-namespace Demo {
-    public partial class ForEachExample {
-        public static int Sum(int[] arr) {
-            int total;
-            total = 0;
-            foreach (var item in arr) total = total + item;
-            return total;
-        }
-    }
-}
