--- tests/GenericBase.cs
@@ -1,6 +1,0 @@
-namespace Demo {
-    public partial class MyList : List<string> {
-        public void Foo() {
-        }
-    }
-}
