--- tests/GenericMethod.cs
@@ -1,6 +1,0 @@
-namespace Demo {
-    public partial class GenClass {
-        public static void DeSerialize<T>(string json, T obj) {
-        }
-    }
-}
