--- tests/GenericStatic.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class GenericStatic {
-        public static void Use() {
-            System.Collections.Generic.List<string>.Sort(null);
-        }
-    }
-}
