--- tests/RepeatUntil.cs
@@ -1,11 +1,0 @@
-namespace Demo {
-    public partial class RepeatExample {
-        public static void LoopIt() {
-            int i;
-            i = 0;
-            do {
-                i = i + 1;
-            } while (!(i > 3));
-        }
-    }
-}
