--- tests/WhileVarInit.cs
@@ -1,8 +1,0 @@
-namespace Demo {
-    public partial class WhileDemo {
-        public static void Test(int[] arr) {
-            int i = 0;
-            while (i <= 2) i = i + arr[0];
-        }
-    }
-}
